Valencio Hoffman - Bezig geweest met de level selector en beetje gespeeld met camera effecten

+++ Game/Super Custom Robot Arena/Assets/Scripts/Utils/GameUtilities.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using Boomlagoon.JSON;
 
 // 
+using System.IO.IsolatedStorage;
 
 public static class GameUtilities {
 	
@@ -21,24 +22,50 @@ public static class GameUtilities {
 	/// </summary>
 	private static string lastLoadProgress = null;
 
+	public static string ReadResource(string Path){
+		TextAsset txtAss = (TextAsset) Resources.Load(Path, typeof(TextAsset));
+		return (txtAss != null) ? txtAss.text : "";
+	}
+	
 	/// <summary>
 	/// Reads the file.
 	/// </summary>
 	/// <returns>The file.</returns>
 	/// <param name="path">Path.</param>
-	public static string ReadFile(string path) {
-		//TextAss declared as public variable and drag dropped the text file in inspector
-		TextAsset TxtAss = (TextAsset) Resources.Load(path, typeof(TextAsset));
-		return (TxtAss != null) ? TxtAss.text : "";
+	public static string ReadFile(string path, string fileName) {
+		try {
+			//TextAss declared as public variable and drag dropped the text file in inspector
+			StreamReader sr;
+			#if UNITY_EDITOR
+			sr = new StreamReader(Application.dataPath + "/Resources/" + path + fileName);
+			#else
+			sr = new StreamReader(Application.persistentDataPath + "/Resources/" + path + fileName);
+			#endif
+			string fileContents = sr.ReadToEnd();
+			sr.Close();
+			return fileContents;
+		}
+		catch (FileNotFoundException ex) {
+			Debug.Log(ex.Message);
+			return "";
+		}				
 	}
 	
 	public static bool CheckFileExists(string path, string fileName){
+		#if UNITY_EDITOR
 		path = Application.dataPath + "/Resources/" + path + fileName;
+		#else
+		path = Application.persistentDataPath + "/Resources/" + path + fileName;
+		#endif
 		return File.Exists(path);
 	}
 	
 	public static bool DeleteFile(string path, string fileName){
+		#if UNITY_EDITOR
 		path = Application.dataPath + "/Resources/" + path + fileName;
+		#else
+		path = Application.persistentDataPath + "/Resources/" + path + fileName;
+		#endif
 		if(File.Exists(path)){
 			File.Delete(path);
 		}
@@ -46,10 +73,42 @@ public static class GameUtilities {
 		return File.Exists(path);
 	}
 	
-	public static void WriteFile(string path, string fileName, string value){		
-		path = Application.dataPath + "/Resources/" + path;
-		File.WriteAllText(path + fileName, value);
+	// Runtime code here
+	public static void WriteFile(string path, string fileName, string value){	
+		#if UNITY_EDITOR
+		WriteEditor(path, fileName, value);
+		#else
+		WriteStandalone(path, fileName, value);
+		#endif
 	}
+	
+	private static void WriteEditor(string path, string fileName, string value){
+		path = Application.dataPath + "/Resources/" + path;
+		try {
+			File.WriteAllText(path + fileName, value);
+		}catch(Exception ex){
+			Debug.Log(ex.Message);
+		}
+	}
+	
+	private static void WriteStandalone(string path, string fileName, string value){
+		string checkPath = Application.persistentDataPath + "/Resources";
+		try {
+			if (!Directory.Exists(checkPath)) {
+				Directory.CreateDirectory(checkPath);
+			}			
+			
+			path = Application.persistentDataPath + "/Resources/" + path;
+			if (!Directory.Exists(path)) {
+				Directory.CreateDirectory(path);
+			}else{
+				File.WriteAllText(path + fileName, value);
+			}
+		}catch (IsolatedStorageException ex) {
+			Debug.Log(ex.Message);
+		}
+	}
+	
 	
 	/// <summary>
 	/// Async loader for the level
